Небольшие исправления

+++ app/controllers/AccountController.php
@@ -40,8 +40,7 @@ class AccountController {
     public function update()
     {
         if ($_FILES['userpic']['size'] != 0) {
-            // TODO при удалении может удалиться дефолтный юзерпик
-            //$this->deleteImageFromDirectory($_SESSION['sid']);
+            $this->deleteImageFromDirectory($_SESSION['sid']);
             $userpic = $this->upload($_FILES['userpic'], $this->userpic_path);
         } else {
             $account = $this->get($_SESSION['sid']);
@@ -57,11 +56,14 @@ class AccountController {
     /**
      * Удаление пользовательского и установка дефолтного юзерпика
      */
-    public function delete_userpic()
+    public function deleteUserpic()
     {
+        $this->deleteImageFromDirectory($_SESSION['sid']);
+
         $data = [];
         foreach ($_POST as $key => $value)
         {
+            // TODO из за json формата, в переменных типа string появляются _
             $data = json_decode($key, true);
         }
 
@@ -74,11 +76,14 @@ class AccountController {
     }
 
     /**
+     * Удаление юзерпика из директории если он не дефолтный
      * @param $id
      */
     public function deleteImageFromDirectory($id)
     {
         $account = $this->get($id);
-        $this->deleteImage($account->userpic);
+        if ($account->userpic != "/userpic_default/userpic.jpg") {
+            $this->deleteImage($account->userpic);
+        }
     }
 }
